chore: remove dead code and relocate test-only scaffolding
A codebase-wide dead-code sweep, cross-checked against `deadcode -test`,
golangci-lint, and the konflate consumer to protect the public API surface.

- Delete genuinely-dead symbols: manifest.ResourceFailedError (superseded by
  DependencyFailedError), schedule.Scheduler.OnDelete, task.Coalescer,
  manifest.filterDependsOn, and the unused internal/testutil status helpers.
- Collapse task's redundant constructors to NewBounded; New/NewUnbounded had
  no production caller (test sites migrate to NewBounded(0)).
- Relocate test-only scaffolding out of the production binary into _test.go:
  manifest.defaultParseDocOptions, values.SliceProvider,
  change.addUngated/addRecursive, and loader.BuildParentIndexForKindWithCache
  (production BuildParentIndexFromPrefixes now carries the contract doc).
- Unexport kustomize.SkipStageDir (no caller outside the package).
- Drop stale signature-verification references left in comments after the
  cosign/PGP removal.

Pure cleanup; render output is unchanged. image.Split and diff.Changes are
kept — the konflate consumer depends on them.
